+++ src/main_window.rs
@@ -118,10 +118,12 @@ impl MainWindowView {
 
     const M_MOVE_ID: &str = "move";
     const M_QUIT_ID: &str = "quit";
+    const M_DEBUG_ID: &str = "debug";
 
     fn create_context_menu() -> anyhow::Result<muda::Menu> {
         muda::Menu::with_items(&[
             &muda::MenuItem::with_id(Self::M_MOVE_ID, "Move", true, None),
+            &muda::CheckMenuItem::with_id(Self::M_DEBUG_ID, "Debugging", true, false, None),
             &muda::MenuItem::with_id(Self::M_QUIT_ID, "Quit", true, None),
         ])
         .map_err(Into::into)
@@ -342,6 +344,7 @@ impl MainWindowView {
 impl EguiView for MainWindowView {
     fn handle_window_event(
         &mut self,
+        _ctx: &egui::Context,
         _event_loop: &ActiveEventLoop,
         event: winit::event::WindowEvent,
     ) -> anyhow::Result<()> {
@@ -359,12 +362,22 @@ impl EguiView for MainWindowView {
 
     fn handle_app_message(
         &mut self,
+        ctx: &egui::Context,
         _event_loop: &ActiveEventLoop,
         message: &AppMessage,
     ) -> anyhow::Result<()> {
         match message {
             AppMessage::UpdateWorkspaces(workspaces) => self.workspaces = workspaces.clone(),
             AppMessage::MenuEvent(e) if e.id() == Self::M_MOVE_ID => self.start_host_dragging()?,
+            AppMessage::MenuEvent(e) if e.id() == Self::M_DEBUG_ID => {
+                let items = self.context_menu.items();
+                let Some(item) = items.iter().find(|i| i.id() == e.id()) else {
+                    return Ok(());
+                };
+                let item = item.as_check_menuitem_unchecked();
+                let is_checked = item.is_checked();
+                ctx.set_debug_on_hover(is_checked);
+            }
             AppMessage::MenuEvent(e) if e.id() == Self::M_QUIT_ID => self.close_host()?,
             AppMessage::SystemSettingsChanged => self.update_system_colors()?,
             _ => {}